Checkstyle adjustments
